src/backend/ml_service/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings
loguru==0.7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# scientific stack (pinned for NumPy 1.26.x compatibility)
numpy==1.26.4
pandas==2.2.2
scipy==1.11.4
scikit-learn==1.3.2
joblib==1.4.2
xgboost==2.1.1
beautifulsoup4==4.12.3
lxml==5.3.0

# multilingual & embeddings
sentence-transformers==3.0.1
google-cloud-translate==2.0.1

# explainability
shap==0.44.1
